EnableWorkDocs applies but does not destroy.
The AWS SDK ModifyWorkspaceCreationProperties auto creates a
WorkDocs Site when the option is toggled to true. This is confirmed
https://docs.aws.amazon.com/workdocs/latest/adminguide/manage-sites.html
by the note under deleting a site.

> If you didn't provide your own directory for Amazon WorkDocs,
> then we created one for you.

This workflow if incompatible with Terraform because

1. The function does not retun the ID or name of the created site.
2. There does not seem to be a way to specify use of a preexisting site.
3. There is no API function to delete the site.
4. While the WorkDocs Site is in existence, attempts to delete the
Workspace Directory will fail.

It seems the only option is to request that AWS improve the API to the
point where I can create a Terraform aws_workspace_workdocs resource.